refact: split filter into smaller components
