+++ src/components/filter/filter.tsx
@@ -1,274 +1,24 @@
 import React, { useState, useEffect } from 'react';
-import { ChevronDown, ChevronRight, X } from 'lucide-react';
+import { X } from 'lucide-react';
 import { useStore } from '@nanostores/react';
 
 import { Button } from '@/components/ui/button';
-import { Checkbox } from '@/components/ui/checkbox';
-import { Field, FieldGroup, FieldLabel } from '@/components/ui/field';
 import {
   isFilterOpen,
   selectedFilterCount,
   $filterState,
   updateFilter,
   resetFilter,
-  FILTER_HIERARCHY,
 } from '@/stores/filterStore';
-import { compareLabel } from '@/lib/array-utils';
-import {
-  type FilterCategory,
-  type AbilityFilter,
-  type StatusDebuff,
-  type EnemyEffect,
-  type DamageBoost,
-  type BeneficialEffect,
-  type EffectUnion,
-} from '@/types/Filter';
 
-// Discriminated union types for our filter structure
-interface BaseFilterItem {
-  id: string;
-  label: string;
-  isOpen?: boolean;
-  isSelected?: boolean;
-}
-
-interface FilterCategoryUI extends BaseFilterItem {
-  type: 'category';
-  id: FilterCategory;
-  children: FilterSubcategoryUI[];
-}
-
-interface FilterSubcategoryUI extends BaseFilterItem {
-  type: 'subcategory';
-  value: AbilityFilter;
-  children: FilterEffectTypeUI[];
-}
-
-interface FilterEffectTypeUI extends BaseFilterItem {
-  type: 'effectType';
-  value: EffectUnion | null;
-  category: FilterCategory | null;
-  subCategory: AbilityFilter;
-}
-
-type FilterItem = FilterCategoryUI | FilterSubcategoryUI | FilterEffectTypeUI;
-
-const formattedEffectLabel: Record<string, string> = {
-  'atk down': 'ATK down',
-  'limit special uses': 'special use limit',
-  'decrease chain multiplier growth rate': 'chain coefficient reduction',
-  atk: 'ATK Boost',
-  hp: 'HP boost',
-  slot: 'slot boost',
-  'heal eot': 'heal EOT',
-  'hp guard': 'HP guard',
-  'final tap atk': 'final tap ATK',
-  'ignited damage boost': 'ignited damage',
-  'def down damage boost': 'def down damage',
-  'poison damage boost': 'poison damage',
-  'percent damage boost': 'percent damage',
-  'delayed damage boost': 'delayed damage',
-  'enemy percent damage': 'percent damage',
-  'enemy threshold damage': 'threshold damage',
-  'enemy def down': 'def down',
-  'enemy paralysis': 'paralysis',
-};
-
-// Helper function to format labels for display
-const formatLabel = (str: string): string => {
-  const formattedStr = formattedEffectLabel[str] || str;
-  return formattedStr
-    .split(/[-_\s]+/)
-    .map((word) => word.charAt(0).toUpperCase() + word.slice(1))
-    .join(' ');
-};
-
-// Build filter structure with proper typing
-const buildFilterStructure = (): FilterCategoryUI[] => {
-  const abilityEntries = Object.keys(FILTER_HIERARCHY.ability) as Array<
-    keyof typeof FILTER_HIERARCHY.ability
-  >;
-  const specialEntries = Object.keys(FILTER_HIERARCHY.special) as Array<
-    keyof typeof FILTER_HIERARCHY.special
-  >;
-
-  return [
-    {
-      id: 'ability',
-      label: 'Ship Ability',
-      type: 'category',
-      isOpen: true,
-      children: abilityEntries.map((key) => {
-        const effectTypes = FILTER_HIERARCHY.ability[key];
-        const subcategoryKey = key as AbilityFilter;
-        const formattedLabel = formatLabel(key as string);
-
-        let children: FilterEffectTypeUI[] = [];
-
-        // Add effect types based on subcategory
-        if (Array.isArray(effectTypes)) {
-          children = (effectTypes as BeneficialEffect[])
-            .map((effectType) => ({
-              id: `ability-${subcategoryKey}-${effectType}`,
-              label: formatLabel(effectType),
-              value: effectType,
-              type: 'effectType' as const,
-              category: 'ability' as const,
-              subCategory: subcategoryKey,
-              isSelected: false,
-            }))
-            .sort(compareLabel);
-        }
-
-        return {
-          id: `ability-${subcategoryKey}`,
-          label: formattedLabel,
-          value: subcategoryKey,
-          type: 'subcategory' as const,
-          isOpen: false,
-          children,
-        };
-      }),
-    },
-    {
-      id: 'special',
-      label: 'Ship Special',
-      type: 'category',
-      isOpen: true,
-      children: specialEntries.map((key) => {
-        const effectTypes = FILTER_HIERARCHY.special[key];
-        const subcategoryKey = key as AbilityFilter;
-        const formattedLabel = formatLabel(key as string);
-
-        let children: FilterEffectTypeUI[] = [];
-
-        // Add effect types based on subcategory
-        if (Array.isArray(effectTypes)) {
-          children = (effectTypes as BeneficialEffect[])
-            .map((effectType) => ({
-              id: `special-${subcategoryKey}-${effectType}`,
-              label: formatLabel(effectType),
-              value: effectType,
-              type: 'effectType' as const,
-              category: 'special' as const,
-              subCategory: subcategoryKey,
-              isSelected: false,
-            }))
-            .sort(compareLabel);
-        }
-
-        return {
-          id: `special-${subcategoryKey}`,
-          label: formattedLabel,
-          value: subcategoryKey,
-          type: 'subcategory' as const,
-          isOpen: false,
-          children,
-        };
-      }),
-    },
-  ];
-};
-
-// Recursive component for rendering filter items
-const FilterItemComponent: React.FC<{
-  item: FilterItem;
-  depth?: number;
-  onToggle?: (id: string) => void;
-  onSelect?: (item: FilterEffectTypeUI | FilterSubcategoryUI) => void;
-}> = ({ item, depth = 0, onToggle, onSelect }) => {
-  const hasChildren =
-    'children' in item && item.children && item.children.length > 0;
-  const isCategory = depth === 0;
-  const isSelectable = depth > 1;
-
-  const handleToggle = () => {
-    if (hasChildren && onToggle) {
-      onToggle(item.id);
-    }
-  };
-
-  const handleSelect = () => {
-    if (isSelectable && onSelect) {
-      if (item.type === 'effectType') {
-        onSelect(item);
-      } else if (item.type === 'subcategory') {
-        onSelect(item);
-      }
-    }
-  };
-
-  const paddingLeft = depth * 8;
-
-  return (
-    <div className="select-none">
-      <div
-        className={`text-sm flex items-center py-2 px-3 hover:bg-gray-100 dark:hover:bg-gray-800 rounded-lg transition-colors ${hasChildren ? 'cursor-pointer' : 'cursor-default'}`}
-        style={{ paddingLeft: `${paddingLeft}px` }}
-        onClick={hasChildren ? handleToggle : undefined}
-      >
-        {/* Chevron icon for expandable items */}
-        {hasChildren && (
-          <span className="mr-2 flex-shrink-0">
-            {item.isOpen ? (
-              <ChevronDown size={16} />
-            ) : (
-              <ChevronRight size={16} />
-            )}
-          </span>
-        )}
-
-        {/* Checkbox for selectable items */}
-        {isSelectable && (
-          <FieldGroup>
-            <Field orientation="horizontal">
-              <Checkbox
-                id={item.id}
-                checked={item.isSelected}
-                onCheckedChange={handleSelect}
-              />
-              <FieldLabel
-                htmlFor={item.id}
-                className="font-normal cursor-pointer"
-              >
-                {item.label}
-              </FieldLabel>
-            </Field>
-          </FieldGroup>
-        )}
-
-        {/* Count badge for categories and subcategories */}
-        {hasChildren && (
-          <>
-            <span
-              className={`flex-grow truncate ${isCategory ? 'font-semibold' : 'font-normal'}`}
-            >
-              {item.label}
-            </span>
-            <span className="ml-2 inline-flex items-center rounded-full bg-gray-100 dark:bg-gray-800 px-2 py-0.5 text-xs font-medium text-gray-700 dark:text-gray-300 flex-shrink-0">
-              {item.children?.length}
-            </span>
-          </>
-        )}
-      </div>
-
-      {/* Render children if open */}
-      {hasChildren && item.isOpen && (
-        <div>
-          {item.children?.map((child) => (
-            <FilterItemComponent
-              key={child.id}
-              item={child}
-              depth={depth + 1}
-              onToggle={onToggle}
-              onSelect={onSelect}
-            />
-          ))}
-        </div>
-      )}
-    </div>
-  );
-};
+import { FilterItemComponent } from './filter-item';
+import { buildFilterStructure } from './utils';
+import type {
+  FilterCategoryUI,
+  FilterEffectTypeUI,
+  FilterItem,
+  FilterSubcategoryUI,
+} from './types';
 
 // Helper function to recursively update categories and their children
 const updateFilterStructure = (
